Split component registry stuff out of index
